Move the ICharacterItem.kt to an api target.
