+++ application/feature/characters/list/api/src/main/java/com/stepango/blockme/feature/characters/list/api/domain/model/ICharacterItem.kt
@@ -1,0 +1,8 @@
+package com.stepango.blockme.feature.characters.list.api.domain.model
+
+interface ICharacterItem {
+    val id: Long
+    val name: String
+    val description: String
+    val imageUrl: String
+}
