+++ application/feature/characters/list/databinding/src/main/java/com/stepango/blockme/feature/characters/list/impl/domain/model/ICharactersListViewModel.kt
@@ -4,6 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.paging.PagedList
 import com.stepango.blockme.core.mvvm.library.lifecycle.SingleLiveData
 import com.stepango.blockme.core.network.library.NetworkState
+import com.stepango.blockme.feature.characters.list.api.domain.model.ICharacterItem
 
 interface ICharactersListViewModel {
     val networkState: LiveData<NetworkState>?
