+++ application/feature/characters/list/databinding/src/main/res/layout/list_item_character.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="character"
-            type="com.stepango.blockme.feature.characters.list.impl.domain.model.ICharacterItem" />
+            type="com.stepango.blockme.feature.characters.list.api.domain.model.ICharacterItem" />
 
         <variable
             name="viewModel"
