+++ application/feature/characters/list/impl/src/main/java/com/stepango/blockme/feature/characters/list/impl/domain/model/CharacterItem.kt
@@ -16,6 +16,8 @@
 
 package com.stepango.blockme.feature.characters.list.impl.domain.model
 
+import com.stepango.blockme.feature.characters.list.api.domain.model.ICharacterItem
+
 data class CharacterItem(
     override val id: Long,
     override val name: String,
