+++ application/feature/characters/list/impl/src/main/java/com/stepango/blockme/feature/characters/list/impl/ui/CharactersListViewModel.kt
@@ -25,7 +25,7 @@ import com.stepango.blockme.core.mvvm.library.lifecycle.SingleLiveData
 import com.stepango.blockme.core.network.library.NetworkState
 import com.stepango.blockme.feature.characters.list.impl.data.datasource.CharactersPageDataSourceFactory
 import com.stepango.blockme.feature.characters.list.impl.data.datasource.PAGE_MAX_ELEMENTS
-import com.stepango.blockme.feature.characters.list.impl.domain.model.ICharacterItem
+import com.stepango.blockme.feature.characters.list.api.domain.model.ICharacterItem
 import com.stepango.blockme.feature.characters.list.impl.domain.model.ICharactersListViewEvent
 import com.stepango.blockme.feature.characters.list.impl.domain.model.ICharactersListViewModel
 import com.stepango.blockme.feature.characters.list.impl.domain.model.ICharactersListViewState
